feat: pre-render catalog/help/legal/whats-new as static, app-free pages
Cold visits to the non-editor content routes used to load the full app
bundle (~2.6 MB: main + Three.js + CodeMirror) just to render a grid of
thumbnails or static text. These four routes are now pre-rendered to
real static HTML at build time — best-in-class for crawlers and instant
for users — and load only the shared 87 KB Tailwind CSS, no app JS.

- New multi-page build: catalog.html / help.html / legal.html /
  whats-new.html. A Vite plugin (src/content/build) injects nav + page
  body + footer (shared chrome) into each at build, sourcing content
  from pure data modules in src/content/data.
- Single source of content: the in-app renderers (src/ui/{catalog,help,
  legal,whatsNew}.ts) now import the SAME data modules, so the static and
  in-app versions can't drift.
- Catalog tiles are baked from public/catalog (name, description, badge,
  category) as <a href=/editor?catalog=…> links; thumbnails hydrate via
  a 759-byte client script — the only JS on any content page.
- Routing: public/_redirects rewrites the clean URLs to *.html on
  Cloudflare; a dev/preview middleware mirrors that so every environment
  matches. The editor still SOFT-renders these pages in-app (no reload)
  via pushState — only real navigations/reloads hit the static pages.
- SEO: per-page <title>/description/canonical/OG/Twitter; /whats-new
  added to the sitemap.

Verified on the production build: each content page loads no
main/three/codemirror/manifold chunk, window.partwright is absent, and
the Tailwind styling resolves.
