+++ catalog.html
@@ -1,0 +1,35 @@
+<!DOCTYPE html>
+<html lang="en" class="dark">
+<head>
+  <meta charset="UTF-8" />
+  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+  <title>Catalog — Premade 3D Models | Partwright</title>
+  <meta name="description" content="Browse Partwright's curated catalog of premade parametric 3D models — JavaScript (manifold-3d), implicit-surface SDF, voxel, OpenSCAD, and solid-CAD (BREP). Open any model in the editor and remix it." />
+  <meta name="author" content="Partwright" />
+
+  <meta property="og:type" content="website" />
+  <meta property="og:site_name" content="Partwright" />
+  <meta property="og:title" content="Catalog — Premade 3D Models | Partwright" />
+  <meta property="og:description" content="A curated gallery of premade parametric 3D models. Open any one in the browser editor and remix it." />
+  <meta property="og:image" content="/og-image.png" />
+  <meta property="og:url" content="/catalog" />
+  <meta name="twitter:card" content="summary_large_image" />
+  <meta name="twitter:title" content="Catalog — Premade 3D Models | Partwright" />
+  <meta name="twitter:description" content="A curated gallery of premade parametric 3D models." />
+  <meta name="twitter:image" content="/og-image.png" />
+  <link rel="canonical" href="/catalog" />
+
+  <link rel="icon" href="/favicon.svg" type="image/svg+xml" />
+  <link rel="icon" href="/favicon-32.png" sizes="32x32" type="image/png" />
+  <link rel="apple-touch-icon" href="/apple-touch-icon.png" />
+  <meta name="theme-color" content="#18181b" />
+  <link rel="manifest" href="/manifest.json" />
+  <link rel="stylesheet" href="/src/style.css" />
+  <link rel="preload" href="/fonts/spacegrotesk-400.woff2" as="font" type="font/woff2" crossorigin="anonymous" />
+  <link rel="preload" href="/fonts/spacegrotesk-600.woff2" as="font" type="font/woff2" crossorigin="anonymous" />
+</head>
+<body class="bg-zinc-900 text-zinc-100 min-h-screen" style="font-family:'Space Grotesk',system-ui,-apple-system,sans-serif">
+  <!--PW-CONTENT-->
+  <script type="module" src="/src/content/catalogEntry.ts"></script>
+</body>
+</html>
