+++ help.html
@@ -1,0 +1,34 @@
+<!DOCTYPE html>
+<html lang="en" class="dark">
+<head>
+  <meta charset="UTF-8" />
+  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+  <title>How Partwright Works — Guide | Partwright</title>
+  <meta name="description" content="A complete guide to Partwright: the code editor, modeling engines (JavaScript/manifold-3d, OpenSCAD, BREP, voxel, SDF), viewport tools, painting, sessions, import/export, and the in-browser AI assistant." />
+  <meta name="author" content="Partwright" />
+
+  <meta property="og:type" content="article" />
+  <meta property="og:site_name" content="Partwright" />
+  <meta property="og:title" content="How Partwright Works — Guide" />
+  <meta property="og:description" content="A complete guide to the editor, modeling engines, viewport, painting, sessions, exports, and the in-browser AI assistant." />
+  <meta property="og:image" content="/og-image.png" />
+  <meta property="og:url" content="/help" />
+  <meta name="twitter:card" content="summary_large_image" />
+  <meta name="twitter:title" content="How Partwright Works — Guide" />
+  <meta name="twitter:description" content="A complete guide to the editor, modeling engines, painting, sessions, exports, and the AI assistant." />
+  <meta name="twitter:image" content="/og-image.png" />
+  <link rel="canonical" href="/help" />
+
+  <link rel="icon" href="/favicon.svg" type="image/svg+xml" />
+  <link rel="icon" href="/favicon-32.png" sizes="32x32" type="image/png" />
+  <link rel="apple-touch-icon" href="/apple-touch-icon.png" />
+  <meta name="theme-color" content="#18181b" />
+  <link rel="manifest" href="/manifest.json" />
+  <link rel="stylesheet" href="/src/style.css" />
+  <link rel="preload" href="/fonts/spacegrotesk-400.woff2" as="font" type="font/woff2" crossorigin="anonymous" />
+  <link rel="preload" href="/fonts/spacegrotesk-600.woff2" as="font" type="font/woff2" crossorigin="anonymous" />
+</head>
+<body class="bg-zinc-900 text-zinc-100 min-h-screen" style="font-family:'Space Grotesk',system-ui,-apple-system,sans-serif">
+  <!--PW-CONTENT-->
+</body>
+</html>
