+++ legal.html
@@ -1,0 +1,34 @@
+<!DOCTYPE html>
+<html lang="en" class="dark">
+<head>
+  <meta charset="UTF-8" />
+  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+  <title>Legal & Privacy — Partwright</title>
+  <meta name="description" content="Partwright's privacy stance (no backend, no telemetry, everything stays in your browser), terms of use and no-warranty notice, and a code-execution disclaimer." />
+  <meta name="author" content="Partwright" />
+
+  <meta property="og:type" content="article" />
+  <meta property="og:site_name" content="Partwright" />
+  <meta property="og:title" content="Legal & Privacy — Partwright" />
+  <meta property="og:description" content="Privacy, terms of use, and a note on running code in your browser." />
+  <meta property="og:image" content="/og-image.png" />
+  <meta property="og:url" content="/legal" />
+  <meta name="twitter:card" content="summary_large_image" />
+  <meta name="twitter:title" content="Legal & Privacy — Partwright" />
+  <meta name="twitter:description" content="Privacy, terms of use, and a note on running code in your browser." />
+  <meta name="twitter:image" content="/og-image.png" />
+  <link rel="canonical" href="/legal" />
+
+  <link rel="icon" href="/favicon.svg" type="image/svg+xml" />
+  <link rel="icon" href="/favicon-32.png" sizes="32x32" type="image/png" />
+  <link rel="apple-touch-icon" href="/apple-touch-icon.png" />
+  <meta name="theme-color" content="#18181b" />
+  <link rel="manifest" href="/manifest.json" />
+  <link rel="stylesheet" href="/src/style.css" />
+  <link rel="preload" href="/fonts/spacegrotesk-400.woff2" as="font" type="font/woff2" crossorigin="anonymous" />
+  <link rel="preload" href="/fonts/spacegrotesk-600.woff2" as="font" type="font/woff2" crossorigin="anonymous" />
+</head>
+<body class="bg-zinc-900 text-zinc-100 min-h-screen" style="font-family:'Space Grotesk',system-ui,-apple-system,sans-serif">
+  <!--PW-CONTENT-->
+</body>
+</html>
