+++ tests/legal.spec.ts
@@ -1,27 +1,25 @@
 import { test, expect } from 'playwright/test';
 
-// Coverage for the /legal route: direct load, footer link, in-page Back, and 404.
+// The /legal route is a pre-rendered, app-free static page (served by
+// _redirects → legal.html on Cloudflare; by the dev/preview middleware
+// locally). Direct navigation gets the static page; the editor still
+// soft-renders an in-app copy from the same content source.
 
-test.beforeEach(async ({ page }) => {
-  // Suppress the first-run guided tour — its backdrop intercepts clicks.
-  await page.addInitScript(() => {
-    try { localStorage.setItem('partwright-tour-completed', '1'); } catch { /* ignore */ }
-  });
-});
-
-test.describe('Legal page', () => {
-  test('renders directly at /legal with the expected sections', async ({ page }) => {
+test.describe('Legal page (static)', () => {
+  test('renders directly at /legal with the expected sections, app-free', async ({ page }) => {
     await page.goto('/legal');
-    await expect(page.locator('#legal-page')).toBeVisible();
     await expect(page.getByRole('heading', { name: 'Legal', level: 1 })).toBeVisible();
-    // Each of the three documented sections is present.
-    await expect(page.locator('#legal-page')).toContainText('Privacy');
-    await expect(page.locator('#legal-page')).toContainText('no warranty');
-    await expect(page.locator('#legal-page')).toContainText('Only run or import code from sources you trust');
-    // License is named explicitly.
-    await expect(page.locator('#legal-page')).toContainText('PolyForm Noncommercial');
-    // The document title reflects the route (the title guard keeps it stable).
-    await expect(page).toHaveTitle(/Legal — Partwright/);
+    const main = page.locator('main');
+    await expect(main).toContainText('Privacy');
+    await expect(main).toContainText('no warranty');
+    await expect(main).toContainText('Only run or import code from sources you trust');
+    await expect(main).toContainText('PolyForm Noncommercial');
+    await expect(page).toHaveTitle(/Legal & Privacy — Partwright/);
+    // Canonical points at the clean route for crawlers.
+    await expect(page.locator('link[rel="canonical"]')).toHaveAttribute('href', /\/legal$/);
+    // The app bundle never boots on this route.
+    expect(await page.evaluate(() => 'partwright' in window)).toBe(false);
+    await expect(page.locator('#ai-panel')).toHaveCount(0);
   });
 
   test('is reachable from the landing-page footer', async ({ page }) => {
@@ -30,25 +28,22 @@ test.describe('Legal page', () => {
     await expect(legalLink).toBeVisible();
     await legalLink.click();
     await expect(page).toHaveURL(/\/legal$/);
-    await expect(page.locator('#legal-page')).toBeVisible();
+    await expect(page.getByRole('heading', { name: 'Legal', level: 1 })).toBeVisible();
   });
 
-  test('Back returns to the editor', async ({ page }) => {
+  test('the nav "Open editor" CTA navigates into the app', async ({ page }) => {
     await page.goto('/legal');
-    await expect(page.locator('#legal-page')).toBeVisible();
-    await page.locator('#legal-page button', { hasText: 'Back' }).click();
-    // The editor's code pane is shown and we're back on the editor route. A
-    // fresh editor visit restores/creates a session, so the URL may gain a
-    // ?session= suffix — match the route with or without it (a strict /editor$
-    // raced the session param being appended right after Back).
-    await expect(page.locator('.cm-content')).toBeVisible();
-    await expect(page).toHaveURL(/\/editor(\?.*)?$/);
+    await page.getByRole('link', { name: /Open editor/i }).click();
+    await expect(page).toHaveURL(/\/editor(\?.*)?$/, { timeout: 30000 });
+    await expect(page.locator('.cm-content')).toBeVisible({ timeout: 30000 });
   });
 
   test('an unknown path still 404s (legal allowlist is exact)', async ({ page }) => {
+    await page.addInitScript(() => {
+      try { localStorage.setItem('partwright-tour-completed', '1'); } catch { /* ignore */ }
+    });
     await page.goto('/legalish');
-    // The 404 page renders, not the legal page.
-    await expect(page.locator('#legal-page')).toHaveCount(0);
+    // The SPA 404 renders (the static legal page is not served for this path).
     await expect(page.locator('body')).toContainText(/not found/i);
   });
 });
