+++ whats-new.html
@@ -1,0 +1,34 @@
+<!DOCTYPE html>
+<html lang="en" class="dark">
+<head>
+  <meta charset="UTF-8" />
+  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+  <title>What's New — Changelog | Partwright</title>
+  <meta name="description" content="A running, human-readable changelog of recently shipped Partwright features — new modeling engines, image relief, sharing, painting, AI controls, and more, newest first." />
+  <meta name="author" content="Partwright" />
+
+  <meta property="og:type" content="article" />
+  <meta property="og:site_name" content="Partwright" />
+  <meta property="og:title" content="What's New — Partwright Changelog" />
+  <meta property="og:description" content="Recently shipped Partwright features, newest first." />
+  <meta property="og:image" content="/og-image.png" />
+  <meta property="og:url" content="/whats-new" />
+  <meta name="twitter:card" content="summary_large_image" />
+  <meta name="twitter:title" content="What's New — Partwright Changelog" />
+  <meta name="twitter:description" content="Recently shipped Partwright features, newest first." />
+  <meta name="twitter:image" content="/og-image.png" />
+  <link rel="canonical" href="/whats-new" />
+
+  <link rel="icon" href="/favicon.svg" type="image/svg+xml" />
+  <link rel="icon" href="/favicon-32.png" sizes="32x32" type="image/png" />
+  <link rel="apple-touch-icon" href="/apple-touch-icon.png" />
+  <meta name="theme-color" content="#18181b" />
+  <link rel="manifest" href="/manifest.json" />
+  <link rel="stylesheet" href="/src/style.css" />
+  <link rel="preload" href="/fonts/spacegrotesk-400.woff2" as="font" type="font/woff2" crossorigin="anonymous" />
+  <link rel="preload" href="/fonts/spacegrotesk-600.woff2" as="font" type="font/woff2" crossorigin="anonymous" />
+</head>
+<body class="bg-zinc-900 text-zinc-100 min-h-screen" style="font-family:'Space Grotesk',system-ui,-apple-system,sans-serif">
+  <!--PW-CONTENT-->
+</body>
+</html>
